Updated GUI
This push is focused on the layout of all visual elements of the GUI that can be displayed independently of game logic. Filler variables have been added to display and test out default values for the game such as each pit containing four stones and the players having a score of zero at the start of each game. The GUI cannot be interacted with at this stage in development, but the next stage will revolve tying the GUI elements to the game logic to make the game visual and playable.