Connects Minimax Algorithm to gui_play.py file
Reverted to commit dea8d08 published on 4/30/25 and carefully implemented 8456f02 GUI connection logic so the game runs the Minimax algorithm. Created a new class in mancala.py called GUIMatch because the original Match was better suited for the console based representation of the board. GUIMatch has more GUI friendly methods and controls turn flow, updates the GUI after each move, and manages interactivity like displaying prompts and enabling/disabling buttons on the player side. MinimaxUIWrapper passes the correct board state to our minimax algorithm, calls for the next move, and can display messages to the GUI display box showing when the AI is thinking and which move it is choosing.